requirements.txt
func-timeout==4.3.5
pandas==1.3.0
numpy==1.24.2
tensorflow== 2.7.0
tensorflow_hub==0.11.0
bert-for-tf2==0.14.7
bert-tensorflow==1.0.4
requests==2.31.0
urllib3==1.26.6
selenium~=4.1.3
scikit-learn==1.0.1
tensorflow-estimator==2.7.0
keras==2.7.0
beautifulsoup4==4.10.0
url-normalize==1.4.3
pymongo==3.12.3
python-dotenv==0.19.0
pika==1.2.0
tldextract~=3.1.2
sklearn-crfsuite~=0.3.6
matplotlib~=3.5.1
openpyxl==3.0.9
gym==0.23.0
webdriver-manager~=3.6.1
protobuf==3.20.1
gin-config==0.5.0
tensorflow-addons==0.15.0
uvicorn==0.15.0
fastapi==0.68.1
natsort==8.2.0